+++ spec/directions/move_solutions_to_answer_key/default_strategy_spec.rb
@@ -53,8 +53,9 @@ RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey::V1 do
       described_class.new.bake(
         chapter: chapter,
         metadata_source: metadata_element,
-        strategy: :american_government,
-        append_to: append_to
+        strategy: :default,
+        append_to: append_to,
+        strategy_options: { classes: %w[review-questions] }
       )
     end