+++ spec/directions/move_solutions_to_answer_key/main_spec.rb
@@ -5,7 +5,7 @@ require 'spec_helper'
 RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey do
   it 'calls v1' do
     expect_any_instance_of(Kitchen::Directions::MoveSolutionsToAnswerKey::V1).to receive(:bake)
-      .with(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
+      .with(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element', strategy_options: {})
     described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
   end
 end